Only serve Snap files from CLI
