+++ packages/snaps-cli/src/webpack/index.ts
@@ -1,3 +1,4 @@
 export * from './compiler';
 export * from './config';
 export * from './plugins';
+export * from './server';
